agent-app/pyproject.toml
[project]
name = "agent-app"
version = "1.0.0"
description = "Asistente Inteligente Manuelita con Memoria y Enrutamiento"
readme = "README.md"
requires-python = ">=3.9"
authors = [
    {name = "Esteban", email = "esteba n@example.com"}
]

dependencies = [
    # Streamlit
    "streamlit>=1.28.0",
    
    # LLM & RAG
    "langchain>=0.1.0",
    "langchain-community>=0.0.10",
    "langchain-core>=0.1.0",
    "langchain-google-genai>=0.0.5",
    "langchain-huggingface>=0.0.1",
    "langchain-text-splitters>=0.0.1",
    
    # Vector & Search
    "chromadb>=0.4.0",
    "sentence-transformers>=2.2.0",
    "rank-bm25>=0.2.2",
    
    # Data Processing
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    
    # Configuration
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    
    # HTTP
    "requests>=2.31.0",
    
    # HTML Processing
    "beautifulsoup4>=4.12.0",
    "html2text>=2020.1.16",
    "lxml>=4.9.0",
    
    # Logging
    "structlog>=23.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"

[tool.black]
line-length = 100
target-version = ["py39"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true

[build-system]
requires = ["setuptools>=65.0"]
build-backend = "setuptools.build_meta"
